refactor: user context 파일 분리
- isAuth 데이터 삭제
- onAuth action 제거
- token 값 추가
