+++ src/contexts/UserContext/initialUserState.js
@@ -1,0 +1,17 @@
+export const initialUserData = {
+  currentUser: {
+    image: null, // 프로필 이미지
+    role: null, // 문구로 사용되는 부분
+    posts: [],
+    likes: [],
+    comments: null,
+    notifications: null,
+    _id: null,
+    fullName: null,
+    email: null,
+    createdAt: null,
+    updatedAt: null,
+  },
+  isLoading: false,
+  token: null,
+}
